Split TTGSnackbar into focused source files
